--- a/src/CompletionProvider.php
+++ b/src/CompletionProvider.php
@@ -201,12 +201,12 @@ class CompletionProvider
                 $this->definitionResolver->resolveExpressionNodeToType($node->dereferencableExpression)
             );
 
-            // The namespaces of the symbol and its parents (eg the implemented interfaces)
-            foreach ($this->expandParentFqns($fqns) as $namespace) {
-                // Collect namespaces definitions
-                foreach ($this->index->getDefinitionsForNamespace($namespace) as $fqn => $def) {
+            // The FQNs of the symbol and its parents (eg the implemented interfaces)
+            foreach ($this->expandParentFqns($fqns) as $parentFqn) {
+                // Collect fqn definitions
+                foreach ($this->index->getDefinitionsForFqn($parentFqn) as $fqn => $def) {
                     // Add the object access operator to only get members of all parents
-                    $prefix = $namespace . '->';
+                    $prefix = $parentFqn . '->';
                     if (substr($fqn, 0, strlen($prefix)) === $prefix && !$def->isMember) {
                         $list->items[] = CompletionItem::fromDefinition($def);
                     }
@@ -231,12 +231,12 @@ class CompletionProvider
                 $classType = $this->definitionResolver->resolveExpressionNodeToType($scoped->scopeResolutionQualifier)
             );
 
-            // The namespaces of the symbol and its parents (eg the implemented interfaces)
-            foreach ($this->expandParentFqns($fqns) as $namespace) {
-                // Collect namespaces definitions
-                foreach ($this->index->getDefinitionsForNamespace($namespace) as $fqn => $def) {
+            // The FQNs of the symbol and its parents (eg the implemented interfaces)
+            foreach ($this->expandParentFqns($fqns) as $parentFqn) {
+                // Collect fqn definitions
+                foreach ($this->index->getDefinitionsForFqn($parentFqn) as $fqn => $def) {
                     // Append :: operator to only get static members of all parents
-                    $prefix = strtolower($namespace . '::');
+                    $prefix = strtolower($parentFqn . '::');
                     if (substr(strtolower($fqn), 0, strlen($prefix)) === $prefix && !$def->isMember) {
                         $list->items[] = CompletionItem::fromDefinition($def);
                     }