--- a/src/CompletionProvider.php
+++ b/src/CompletionProvider.php
@@ -207,7 +207,7 @@ class CompletionProvider
                 foreach ($this->index->getDefinitionsForFqn($parentFqn) as $fqn => $def) {
                     // Add the object access operator to only get members of all parents
                     $prefix = $parentFqn . '->';
-                    if (substr($fqn, 0, strlen($prefix)) === $prefix && !$def->isMember) {
+                    if (substr($fqn, 0, strlen($prefix)) === $prefix && $def->isMember) {
                         $list->items[] = CompletionItem::fromDefinition($def);
                     }
                 }
@@ -237,7 +237,7 @@ class CompletionProvider
                 foreach ($this->index->getDefinitionsForFqn($parentFqn) as $fqn => $def) {
                     // Append :: operator to only get static members of all parents
                     $prefix = strtolower($parentFqn . '::');
-                    if (substr(strtolower($fqn), 0, strlen($prefix)) === $prefix && !$def->isMember) {
+                    if (substr(strtolower($fqn), 0, strlen($prefix)) === $prefix && $def->isMember) {
                         $list->items[] = CompletionItem::fromDefinition($def);
                     }
                 }
@@ -301,25 +301,29 @@ class CompletionProvider
                 }
             }
 
-            // Suggest global symbols that either
+            // Suggest global (ie non member) symbols that either
             //  - start with the current namespace + prefix, if the Name node is not fully qualified
             //  - start with just the prefix, if the Name node is fully qualified
-            foreach ($this->index->getGlobalDefinitions() as $fqn => $def) {
+            foreach ($this->index->getDefinitions() as $fqn => $def) {
 
                 $fqnStartsWithPrefix = substr($fqn, 0, $prefixLen) === $prefix;
 
                 if (
-                    !$prefix
-                    || (
-                        // Either not qualified, but a matching prefix with global fallback
-                        ($def->roamed && !$isQualified && $fqnStartsWithPrefix)
-                        // Or not in a namespace or a fully qualified name or AND matching the prefix
-                        || ((!$namespaceNode || $isFullyQualified) && $fqnStartsWithPrefix)
-                        // Or in a namespace, not fully qualified and matching the prefix + current namespace
+                    // Exclude methods, properties etc.
+                    !$def->isMember
+                    && (
+                        !$prefix
                         || (
-                            $namespaceNode
-                            && !$isFullyQualified
-                            && substr($fqn, 0, $namespacedPrefixLen) === $namespacedPrefix
+                            // Either not qualified, but a matching prefix with global fallback
+                            ($def->roamed && !$isQualified && $fqnStartsWithPrefix)
+                            // Or not in a namespace or a fully qualified name or AND matching the prefix
+                            || ((!$namespaceNode || $isFullyQualified) && $fqnStartsWithPrefix)
+                            // Or in a namespace, not fully qualified and matching the prefix + current namespace
+                            || (
+                                $namespaceNode
+                                && !$isFullyQualified
+                                && substr($fqn, 0, $namespacedPrefixLen) === $namespacedPrefix
+                            )
                         )
                     )
                     // Only suggest classes for `new`