… src/Index/Index.php
… src/Index/ReadableIndex.php
… src/Indexer.php
… src/Server/TextDocument.php
… tests/Server/TextDocument/CompletionTest.php
… tests/Server/Workspace/SymbolTest.php