@nicolasmure 2017-08-06
I've added an other commit to ceate an index representation of the global definitions (reduces the number of checks).
I think that PHP does doe make copies but store references to the definition objects.
As you can see [here](https://github.com/felixfbecker/php-language-server/pull/451/commits/f1419c1701eac825c84264e49ea2d509c9bce48c#diff-18eb38922fb7235266f15163e2b050edR238), I checked the memory consumption usage with `memory_get_usage(true)`.

However, I have some other memory consumption metrics : 
on the current master branch : 
![screenshot from 2017-08-07 00-56-30](https://user-images.githubusercontent.com/4362252/29007864-74c57918-7b0c-11e7-8a04-519547c455b2.png)

on this branch : 
![screenshot from 2017-08-07 00-58-05](https://user-images.githubusercontent.com/4362252/29007874-d5262884-7b0c-11e7-9a3d-dbd50cb29f04.png)