@nicolasmure 2017-08-07
I saw that you [had a discussion](https://github.com/felixfbecker/php-language-server/issues/108) with the maintainer of [php-integrator](https://github.com/php-integrator/atom-base), it could be a good idea to store the indexes in a sqlite DB too instead of in-memory and serialized files (will reduce the RAM consumption IMHO).