@felixfbecker 2017-08-08
Could you try always using the tree? I could imagine it is not significantly slower than the map