@nicolasmure 2017-08-09
Hello @felixfbecker , I'm not sure to get what you're saying :thinking: 
Do you mean to have an index like this : 
```php
$this->definitions['Symfony\Component\HttpFoundation\Request'] => [
    'Symfony\Component\HttpFoundation\Request->getLocale()' => $definition
]
```
and to get rid of the `namespaceDefinitions` index ? It would result in a BC break because the serialized index cache no longer be the same.