@felixfbecker 2017-08-09
I mean to get rid of `$definitions` and always use `namespaceDefinitions`.
Changing the cache format is not a BC break, you can simply increase the cache format version.