@nicolasmure 2017-08-09
OK @felixfbecker , I added 3 commits to save memory, here are the consumptions (on my testing project):

| Master | Before these 3 commits | After |
| ------------- | ------------- | ------------- |
| 518Mb  | 529Mb  | 522Mb  |

I also yield indexec definitions to save memory instead of copying them in an array.