@roblourens 2017-08-09
Do they have an impact on the time? I remember from looking at these earlier, that they can add a surprising overhead.