@nicolasmure 2017-08-09
@roblourens , here are some time records grabbed for the [accessor (`->`) operator](https://github.com/nicolasmure/php-language-server/blob/6ba3d6c0c52cd9177d68a3701ef81159dda13aa9/src/CompletionProvider.php#L246-L254) : 

Before these 3 commmits : 
```
0.00016880035400391s
0.00014901161193848s
0.00014591217041016s
0.00020003318786621s
0.00025105476379395s
0.00034809112548828s
0.00015783309936523s
0.00014805793762207s
0.00018405914306641s
0.00019478797912598s
0.00015902519226074s
```

after : 
```
0.0003349781036377s
0.00020289421081543s
0.00030088424682617s
0.00017595291137695s
0.00033283233642578s
0.00017690658569336s
0.0005340576171875s
0.00024986267089844s
0.0003349781036377s
0.00018405914306641s
0.00019097328186035s
```