@felixfbecker 2017-08-10
Generators should actually not decrease memory usage because PHP variables are copy-on-write. But it will reduce time because you save an iteration.