@staabm 2017-08-10
Iterator also can safe memory when you no longer need to build whole arrays but work item per item (stream like)