@felixfbecker 2017-08-10
That's true if the source is an IO stream, but here the source is an already in-memory array.