@nicolasmure 2017-08-10
@staabm yes and it is definitively the case here : we're just looping over it, using each item one by one (no need to know how many items are in the iterator, nor to access to the ones next to the currently used).

@felixfbecker as the index used by the `LanguageServer` is a `AbstractAggregateIndex`, it makes sense IMO to use generators to 'stream' definitions from the aggregated indexes.