@robclancy 2017-08-10
Master: 454MiB
Before the memory saving commits here: 466MiB
Now: 392 MiB