@nicolasmure 2017-08-10
@robclancy yes, make sure that the cache has finished to be populated too before making any mesurements. The last commits have changed the cache version so it has the server has to rebuild all the cache items, so memory consumption will start from 0 to x until it has completed the project parsing.

You can see when it has finished to built by wathcing at CPU consumption and also by listing the cache directory by last modification date : 
```bash
$ watch -n 0.5 ls -c ~/.phpls/
```