@robclancy 2017-08-10
I'm getting my numbers from the output in vscode after the indexing is done.