@nicolasmure 2017-10-05
I've added a commit to correct the terminology (as you [pointed out](https://github.com/felixfbecker/php-language-server/pull/451#discussion_r141475428) @felixfbecker ).
The serialized cache key has changed (`namespaceDefinitions` => `fqnDefinitions`) so be sure to remove the previous cache files generated on this branch before using this lastest commit.