@nicolasmure 2017-11-14
So I'm now using the single (and multi level) definition index array, but even if phpunit tests are passing, there is an error with what is proposed as completion items. I have this problem on nvim and vscode, they're both sending the same request to the server so it does not appear to be a client error.

The problem is that the resolved expression (ie the left hand side of the cursor) is not good. For instance, if you want to see what are the completion possibilities for the `$definition` var of [this function](https://github.com/nicolasmure/php-language-server/blob/e162d94e14114601db77d5ae8943b43c1cbf88ac/src/Index/Index.php#L164), you can add a new line at the beginning of this function, such as 
```php
$definition->get
```
but you'll see that the proposed completion items are not the good one (thay don't belong to the $definition object). Actually, in this exemple they belong to the `$this` object.

I think there might be a n issue with the [node resolution](https://github.com/nicolasmure/php-language-server/blob/e162d94e14114601db77d5ae8943b43c1cbf88ac/src/CompletionProvider.php#L130) ([see here](https://github.com/nicolasmure/php-language-server/blob/e162d94e14114601db77d5ae8943b43c1cbf88ac/src/PhpDocument.php#L215), which use the [microsoft/tolerant-php-parser](https://github.com/Microsoft/tolerant-php-parser)).
The strange point is that I don't have this problem on the master branch :/

Any hints on this ?