@felixfbecker 2017-11-15
@nicolasmure if you say there is a newline, why can't we strip it? What is different on master? If you push your WIP I can also take a look