@felixfbecker 2017-11-15
@nicolasmure I cannot reproduce your issue:

![image](https://user-images.githubusercontent.com/10532611/32821791-12f6ec54-c98a-11e7-8e3a-056f8d604d0a.png)