@felixfbecker 2017-11-15
@nicolasmure I cannot reproduce your issue:

![image](https://user-images.githubusercontent.com/10532611/32821816-4923a2a4-c98a-11e7-8168-f4898aae4907.png)