@felixfbecker 2017-11-15
Other than https://github.com/felixfbecker/php-language-server/pull/451#discussion_r151031720 I think this would be fine to merge