@Declspeck 2018-02-03
I made the requested change in [my branch](https://github.com/Declspeck/php-language-server/tree/autocomplet-speedup).

Turns out I can't push commits to this PR, and I wouldn't really like to open a new PR because that would make it seem like I did all the hard work, not @nicolasmure. 

Benchmarks before and after (`benchmarks/completion.php`, as modified in my branch):

Before:
```
Getting completion
Time ($this->|): 0.0012650489807129s
90 completion items
Time (new|): 0.095192909240723s
1 completion items

```

After:
```
Getting completion
Time ($this->|): 0.0012638568878174s
90 completion items
Time (new|): 0.00032806396484375s
1 completion items
```